Separate head visualization's classes in its own documents
