+++ Zilon.Core/CDT.LAST.MonoGameClient/ViewModels/MainScene/BodyPart.cs
@@ -1,11 +1,10 @@
-﻿using System;
-
+﻿
 using Microsoft.Xna.Framework.Graphics;
 
 namespace CDT.LAST.MonoGameClient.ViewModels.MainScene
 {
     /// <summary>
-    /// Helper structure to keep texture and metadata forbody visualization.
+    /// Helper structure to keep texture and metadata for body visualization.
     /// </summary>
     public record BodyPart
     {
@@ -18,25 +17,5 @@ namespace CDT.LAST.MonoGameClient.ViewModels.MainScene
         public Texture2D Texture { get; }
 
         public BodyPartType Type { get; }
-    }
-
-    public record HeadPart
-    {
-        public HeadPart(HeadPartType type, Texture2D texture)
-        {
-            Texture = texture ?? throw new ArgumentNullException(nameof(texture));
-            Type = type;
-        }
-
-        public Texture2D Texture { get; }
-
-        public HeadPartType Type { get; }
-    }
-
-    public enum HeadPartType
-    {
-        Undefined,
-        Base,
-        Inside
     }
 }
