+++ Zilon.Core/CDT.LAST.MonoGameClient/ViewModels/MainScene/HeadPartType.cs
@@ -1,0 +1,17 @@
+﻿namespace CDT.LAST.MonoGameClient.ViewModels.MainScene
+{
+    /// <summary>
+    /// The part of person head visualization available on the client.
+    /// </summary>
+    public enum HeadPartType
+    {
+        Undefined,
+        Base,
+
+        /// <summary>
+        /// Visualization inside of the head.
+        /// Used in a closed helmets to shade the person face inside a helmet.
+        /// </summary>
+        Inside
+    }
+}
